refactoring for codeclimate
