+++ app/src/main/java/ch/epfl/sweng/radius/browseProfiles/CustomAdapter.java
@@ -1,7 +1,6 @@
 package ch.epfl.sweng.radius.browseProfiles;
 
 import android.content.Context;
-import android.content.Intent;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
@@ -57,16 +56,7 @@ public class CustomAdapter extends ArrayAdapter<ChatListItem> {
         final int clickedPic = chatListItem.getImage();
         final String clickedName = chatListItem.getName();
 
-        imageView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(context, BrowseProfilesActivity.class);
-                intent.putExtra("Clicked Picture", clickedPic);
-                intent.putExtra("Clicked Name", clickedName);
-                context.startActivity(intent);
-            }
-        });
-
+        CustomListener.setCustomOnClick(view, imageView, clickedPic, clickedName, context);
         return view;
     }
 }
